+++ CHANGELOG.md
@@ -1,6 +1,21 @@
-# [1.0.0-beta.8](http://bitbucket.org/uclaucomm/ucla-bruin-components/compare/v1.0.0-beta.7...v1.0.0-beta.8) (2021-01-22)
+# [1.0.0-dev-beta.11](http://bitbucket.org/uclaucomm/ucla-bruin-components/compare/v1.0.0-dev-beta.10...v1.0.0-dev-beta.11) (2021-04-06)
 
 
 ### Features
 
-* **new components and fixes:** adding Alert and Tabs Components, Fixing style lint errors ([e95d77f](http://bitbucket.org/uclaucomm/ucla-bruin-components/commits/e95d77f51b785c38ff0b95c690aacb0fd6cc0bf2))
+* **accordion, primary navigation:** added Accordion and Primary Navigation Component ([3bf151d](http://bitbucket.org/uclaucomm/ucla-bruin-components/commits/3bf151dcaacf7b4fab1682e97b26bf7590e2823d)), closes [#6](http://bitbucket.org/uclaucomm/ucla-bruin-components/issue/6) [#8](http://bitbucket.org/uclaucomm/ucla-bruin-components/issue/8)
+* **primary navigation, grid:** primary Nav Escape and Grid Spacing ([60e36bd](http://bitbucket.org/uclaucomm/ucla-bruin-components/commits/60e36bd680ff6def1b14118b9d09632176527a3b))
+
+# [1.0.0-dev-beta.10](http://bitbucket.org/uclaucomm/ucla-bruin-components/compare/v1.0.0-dev-beta.9...v1.0.0-dev-beta.10) (2021-03-16)
+
+
+### Features
+
+* **footer and email:** including footer and email sign up component ([32fb5bf](http://bitbucket.org/uclaucomm/ucla-bruin-components/commits/32fb5bfc1b41184934bea37f7c3347ac0c7d67d2))
+
+# [1.0.0-dev-beta.9](http://bitbucket.org/uclaucomm/ucla-bruin-components/compare/v1.0.0-dev-beta.8...v1.0.0-dev-beta.9) (2021-02-12)
+
+
+### Features
+
+* **tile component:** including tile component ([4853bc0](http://bitbucket.org/uclaucomm/ucla-bruin-components/commits/4853bc0d8a26f31f3305b567bae22393a75504a2))